feat: added cursor hover and element placement
